feat(engine-core): ✨ feat: Add edit button to Object Layer Management table
Introduces an EditButtonRenderer custom cell renderer in object-layer.management.js to display an edit icon button in the management table. This button:

Reloads the ObjectLayerEngineModal if its open, or opens it otherwise.

Sets the current path with the selected objects ID as a query parameter.

Updates ObjectLayerEngineModal with a Reload function to allow content refresh when already open.