Add Packer CLI installation and use spawnSync for build
Adds a new option --install-packer to the baremetal CLI command to
facilitate the installation of the Packer CLI tool using a setup script.

Refactors the Packer image build process to use spawnSync for packer
init and packer build commands. This ensures a blocking execution and
allows inheriting I/O streams for better user visibility during the
build.

A headless variable is also introduced in the Rocky9 Packer
configuration to control the QEMU source's headless mode.