Add support for primary MongoDB pod ensure and cluster flags in db commands
This commit introduces the --primary-pod-ensure option to the db
subcommands in the CLI.

If a pod-name is provided for --primary-pod-ensure, the system will
check for the existence of the MongoDB primary pod associated with that
name. If the primary pod is not found, a command is executed to ensure
the MongoDB cluster setup, utilizing the provided cluster flags (--dev,
--k3s, --kubeadm, --kind) if present.

Changes include:

Added primaryPodEnsure, dev, k3s, kubeadm, and kind options to
UnderpostDB constructor in src/cli/db.js.

Implemented logic in UnderpostDB to check and ensure the MongoDB primary
pod if options.primaryPodEnsure is set, using the provided cluster flags
to determine the correct setup command.

Added the corresponding command line options (--primary-pod-ensure,
--dev, --kubeadm, --kind, --k3s) to the db program in src/cli/index.js.